+++ src/Hexapod_Serial.h
@@ -30,7 +30,7 @@ class Hexapod_Serial
 private:
     char buffer[MAX_BUF];          // where we store the message until we get a ';'
     int sofar = 0;                 // how much is in the buffer
-    int x_axis, y_axis, z_axis, roll, pitch, yaw; // positions
+    double x_axis, y_axis, z_axis, roll, pitch, yaw; // positions
     char mode_abs = 1;             // absolute mode?
     long line_number = 0;
     char str[20];
@@ -39,7 +39,7 @@ private:
     void processCommand();
     double parseNumber(const char code, double val);
     void output(const char code, double val);
-    void scaleAndConstrain(int *value, int range);
+    void scaleAndConstrain(double *value, int range);
     void logAxes(void);
 
 public:
